+++ tests/test_rdfa.py
@@ -248,6 +248,48 @@ def test_a_nested_subject_is_the_value_of_its_property():
     ]
 
 
+def test_a_subject_whose_property_has_a_literal_value_holds_that_property():
+    # Drupal 7 writes a node's author and tags this way. With `datatype` or
+    # `content` on it, a `typeof` element's property is the new subject's own,
+    # not a link to it from the subject around it: RDFa Core's processing
+    # rules, and the W3C RDFa suite's test 0317. The names were lost.
+    doc = load(
+        '<article typeof="sioc:Item">'
+        '<span property="dc:title" content="Antimodes"></span>'
+        '<span rel="sioc:has_creator">'
+        '<span typeof="sioc:UserAccount" property="foaf:name" datatype="">malte'
+        "</span></span>"
+        '<a href="/tags/islam" typeof="skos:Concept" '
+        'property="rdfs:label skos:prefLabel" datatype="">Islam</a>'
+        "</article>"
+    )
+
+    sioc, foaf, skos, rdfs = (
+        "http://rdfs.org/sioc/ns#",
+        "http://xmlns.com/foaf/0.1/",
+        "http://www.w3.org/2004/02/skos/core#",
+        "http://www.w3.org/2000/01/rdf-schema#",
+    )
+    assert read_rdfa(doc) == [
+        {"@type": sioc + "Item", "http://purl.org/dc/terms/title": "Antimodes"},
+        {"@type": sioc + "UserAccount", foaf + "name": "malte"},
+        {
+            "@type": skos + "Concept",
+            rdfs + "label": "Islam",
+            skos + "prefLabel": "Islam",
+        },
+    ]
+
+
+def test_a_datatype_asks_for_the_words_not_the_address():
+    doc = load(
+        '<div typeof="Article">'
+        '<a property="keywords" datatype="" href="/tags/brakes">brakes</a></div>'
+    )
+
+    assert read_rdfa(doc)[0]["keywords"] == "brakes"
+
+
 def test_a_subject_declaring_nothing_at_all_is_not_a_record():
     doc = load('<div typeof=""><span>Brake pad set</span></div>')